# apps/server-go/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Technews HTTP API compatibility manifest
  version: 2.0.0
  description: >-
    Success-only description of the 32 canonical operations in the reviewed
    schema-v2 synthetic Node contract capture. These operations capture reviewed
    primary success responses only. The generated fixture remains the compatibility
    authority for concrete values, optionality, and nullability. Negative and error
    scenarios, including the typed 401 observation and multipart errors, are tracked
    but deferred; each must be added by a future vertical slice before error parity
    is claimed.
servers:
  - url: /
paths:
...
  /api/articles:
    get:
      operationId: articles.list
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Paginated articles
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArticlePage'
  /api/articles/trending:
    get:
      operationId: articles.trending
      parameters:
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Trending articles
          content:
            application/json:
              schema:
                type: object
                properties:
                  articles:
                    type: array
                    items:
                      $ref: '#/components/schemas/Article'
  /api/articles/{slug}:
    get:
      operationId: articles.getBySlug
      parameters:
        - $ref: '#/components/parameters/Slug'
      responses:
        '200': &articleResponse
          description: Article
          content:
            application/json:
              schema:
                type: object
                properties:
                  article:
                    $ref: '#/components/schemas/Article'
  /api/articles/id/{id}:
    get:
      operationId: articles.getById
      parameters:
        - $ref: '#/components/parameters/ID'
      responses:
        '200': *articleResponse
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    cronBearer:
      type: http
      scheme: bearer
  parameters:
    ID:
      name: id
      in: path
      required: true
      schema: {type: integer}
    Slug:
      name: slug
      in: path
      required: true
      schema: {type: string}
    Page:
      name: page
      in: query
      schema: {type: integer}
    Limit:
      name: limit
      in: query
      schema: {type: integer}
...
  schemas:
...
    Pagination:
      type: object
      properties:
        total: {type: integer}
        page: {type: integer}
        totalPages: {type: integer}
    Category:
      type: object
      properties:
        id: {type: integer}
        name: {type: string}
        slug: {type: string}
        description: {type: string}
        color: {type: string}
        article_count: {type: integer}
...
    Author:
      type: object
      properties:
        id: {type: integer}
        name: {type: string}
        email: {type: string, format: email}
        avatar: {type: [string, 'null']}
        bio: {type: [string, 'null']}
        role: {type: string}
    Article:
      type: object
      properties:
        id: {type: integer}
        title: {type: string}
        slug: {type: string}
        excerpt: {type: string}
        content: {type: string}
        featured_image: {type: [string, 'null']}
        category_id: {type: integer}
        author_id: {type: integer}
        status: {type: string}
        published_at: {type: [string, 'null']}
        meta_title: {type: [string, 'null']}
        meta_description: {type: [string, 'null']}
        # source and source_url appear only in dashboard (admin) responses;
        # the public /api/articles routes never include them.
        source: {type: [string, 'null']}
        source_url: {type: [string, 'null']}
        view_count: {type: integer}
        # Public single-article reads only, when the article has a summary.
        tldr: {type: array, items: {type: string}}
        why_it_matters: {type: string}
        created_at: {type: string}
        updated_at: {type: string}
        category:
          $ref: '#/components/schemas/Category'
        author:
          $ref: '#/components/schemas/Author'
...
    ArticlePage:
      allOf:
        - $ref: '#/components/schemas/Pagination'
        - type: object
          properties:
            articles:
              type: array
              items:
                $ref: '#/components/schemas/Article'
...